Add redirect so tapping inbox tab defaults to open "Inbox (all)"
Resolves #458
