+++ src/pages/inbox/BoxesPage.tsx
@@ -22,6 +22,7 @@ import { getInboxCounts } from "../../features/inbox/inboxSlice";
 import { MouseEvent, useContext, useRef } from "react";
 import { PageContext } from "../../features/auth/PageContext";
 import { useSetActivePage } from "../../features/auth/AppContext";
+import BoxesRedirectBootstrapper from "../../features/inbox/BoxesRedirectBootstrapper";
 
 export default function BoxesPage() {
   const pageRef = useRef<HTMLElement>(null);
@@ -40,67 +41,73 @@ export default function BoxesPage() {
   }
 
   return (
-    <IonPage ref={pageRef} className="grey-bg">
-      <IonHeader>
-        <IonToolbar>
-          <IonTitle>Boxes</IonTitle>
-        </IonToolbar>
-      </IonHeader>
-      <AppContent scrollY fullscreen>
-        <IonHeader collapse="condense">
+    <>
+      <BoxesRedirectBootstrapper />
+      <IonPage ref={pageRef} className="grey-bg">
+        <IonHeader>
           <IonToolbar>
-            <IonTitle size="large">Boxes</IonTitle>
+            <IonTitle>Boxes</IonTitle>
           </IonToolbar>
         </IonHeader>
+        <AppContent scrollY fullscreen>
+          <IonHeader collapse="condense">
+            <IonToolbar>
+              <IonTitle size="large">Boxes</IonTitle>
+            </IonToolbar>
+          </IonHeader>
 
-        <IonList inset color="primary">
-          <InsetIonItem routerLink="/inbox/all" onClick={interceptIfLoggedOut}>
-            <IonIcon icon={fileTray} color="primary" />
-            <SettingLabel>Inbox (All)</SettingLabel>
-          </InsetIonItem>
-          <InsetIonItem
-            routerLink="/inbox/unread"
-            onClick={interceptIfLoggedOut}
-          >
-            <IonIcon icon={ellipseOutline} color="primary" />
-            <SettingLabel>Unread</SettingLabel>
-          </InsetIonItem>
-        </IonList>
+          <IonList inset color="primary">
+            <InsetIonItem
+              routerLink="/inbox/all"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={fileTray} color="primary" />
+              <SettingLabel>Inbox (All)</SettingLabel>
+            </InsetIonItem>
+            <InsetIonItem
+              routerLink="/inbox/unread"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={ellipseOutline} color="primary" />
+              <SettingLabel>Unread</SettingLabel>
+            </InsetIonItem>
+          </IonList>
 
-        <IonList inset color="primary">
-          <InsetIonItem
-            routerLink="/inbox/post-replies"
-            onClick={interceptIfLoggedOut}
-          >
-            <IonIcon icon={albumsOutline} color="primary" />
-            <SettingLabel>Post Replies</SettingLabel>
-          </InsetIonItem>
-          <InsetIonItem
-            routerLink="/inbox/comment-replies"
-            onClick={interceptIfLoggedOut}
-          >
-            <IonIcon icon={chatbubbleOutline} color="primary" />
-            <SettingLabel>Comment Replies</SettingLabel>
-          </InsetIonItem>
-          <InsetIonItem
-            routerLink="/inbox/mentions"
-            onClick={interceptIfLoggedOut}
-          >
-            <IonIcon icon={personCircleOutline} color="primary" />
-            <SettingLabel>Mentions</SettingLabel>
-          </InsetIonItem>
-        </IonList>
+          <IonList inset color="primary">
+            <InsetIonItem
+              routerLink="/inbox/post-replies"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={albumsOutline} color="primary" />
+              <SettingLabel>Post Replies</SettingLabel>
+            </InsetIonItem>
+            <InsetIonItem
+              routerLink="/inbox/comment-replies"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={chatbubbleOutline} color="primary" />
+              <SettingLabel>Comment Replies</SettingLabel>
+            </InsetIonItem>
+            <InsetIonItem
+              routerLink="/inbox/mentions"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={personCircleOutline} color="primary" />
+              <SettingLabel>Mentions</SettingLabel>
+            </InsetIonItem>
+          </IonList>
 
-        <IonList inset color="primary">
-          <InsetIonItem
-            routerLink="/inbox/messages"
-            onClick={interceptIfLoggedOut}
-          >
-            <IonIcon icon={mail} color="primary" />
-            <SettingLabel>Messages</SettingLabel>
-          </InsetIonItem>
-        </IonList>
-      </AppContent>
-    </IonPage>
+          <IonList inset color="primary">
+            <InsetIonItem
+              routerLink="/inbox/messages"
+              onClick={interceptIfLoggedOut}
+            >
+              <IonIcon icon={mail} color="primary" />
+              <SettingLabel>Messages</SettingLabel>
+            </InsetIonItem>
+          </IonList>
+        </AppContent>
+      </IonPage>
+    </>
   );
 }
