+++ src/pages/posts/CommunitiesPage.tsx
@@ -1,4 +1,4 @@
-import { useRef } from "react";
+import { memo, useRef } from "react";
 import CommunitiesList from "../../features/community/list/CommunitiesList";
 import { useSetActivePage } from "../../features/auth/AppContext";
 import {
@@ -13,8 +13,19 @@ import FeedContent from "../shared/FeedContent";
 import { useParams } from "react-router";
 import CommunitiesListRedirectBootstrapper from "../../features/community/list/CommunitiesListRedirectBootstrapper";
 
+interface CommunitiesPageParams {
+  actor: string;
+}
+
 export default function CommunitiesPage() {
-  const { actor } = useParams<{ actor: string }>();
+  const { actor } = useParams<CommunitiesPageParams>();
+
+  return <CommunitiesPageContent actor={actor} />;
+}
+
+const CommunitiesPageContent = memo(function CommunitiesPageContent({
+  actor,
+}: CommunitiesPageParams) {
   const pageRef = useRef<HTMLElement>(null);
 
   useSetActivePage(pageRef);
@@ -37,4 +48,4 @@ export default function CommunitiesPage() {
       </IonPage>
     </>
   );
-}
+});
